pyproject.toml
[project]
name = "whatsapp-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.11",
    "llama-index-llms-groq>=0.3.1",
    "pydantic-settings>=2.8.1",
    "uvicorn>=0.34.0",
    "vectara-agentic>=0.2.4",
]
